+++ src/plopp/data/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
+
+# flake8: noqa F401
+
+from .generated import variable, data_array, dataset, scatter_data
